+++ mink/tasks/__init__.py
@@ -4,9 +4,11 @@ from .com_task import ComTask
 from .damping_task import DampingTask
 from .equality_constraint_task import EqualityConstraintTask
 from .frame_task import FrameTask
+from .kinetic_energy_regularization_task import KineticEnergyRegularizationTask
+from .l2_regularization_task import L2RegularizationTask
 from .posture_task import PostureTask
 from .relative_frame_task import RelativeFrameTask
-from .task import Objective, Task
+from .task import Objective, RegularizationTask, Task
 
 __all__ = (
     "ComTask",
@@ -15,6 +17,9 @@ __all__ = (
     "DampingTask",
     "PostureTask",
     "RelativeFrameTask",
+    "RegularizationTask",
     "Task",
     "EqualityConstraintTask",
+    "KineticEnergyRegularizationTask",
+    "L2RegularizationTask",
 )